+++ packages/flutter_riverpod/lib/flutter_riverpod.dart
@@ -5,9 +5,20 @@ export 'src/change_notifier_provider.dart'
         ChangeNotifierProvider,
         AutoDisposeChangeNotifierProvider,
         AutoDisposeChangeNotifierProviderFamily,
-        ChangeNotifierProviderFamily;
+        ChangeNotifierProviderFamily,
+        AutoDisposeChangeNotifierProviderRef,
+        ChangeNotifierProviderRef,
+        XAutoDisposeChangeNotifierFamily,
+        XChangeNotifierFamily;
 export 'src/consumer.dart'
-    show Consumer, ConsumerWidget, ConsumerBuilder, WidgetRef;
+    show
+        Consumer,
+        ConsumerWidget,
+        ConsumerBuilder,
+        WidgetRef,
+        ConsumerState,
+        ConsumerStatefulWidget,
+        ConsumerStatefulElement;
 export 'src/framework.dart' show ProviderScope, UncontrolledProviderScope;
 // ignore: deprecated_member_use_from_same_package
 export 'src/provider_listener.dart' show ProviderListener, OnProviderChange;